+++ oci/core_security_lists_data_source.go
@@ -33,7 +33,7 @@ func CoreSecurityListsDataSource() *schema.Resource {
 			},
 			"vcn_id": {
 				Type:     schema.TypeString,
-				Required: true,
+				Optional: true,
 			},
 			"security_lists": {
 				Type:     schema.TypeList,
@@ -118,7 +118,6 @@ func (s *CoreSecurityListsDataSourceCrud) SetData() error {
 	for _, r := range s.Res.Items {
 		securityList := map[string]interface{}{
 			"compartment_id": *r.CompartmentId,
-			"vcn_id":         *r.VcnId,
 		}
 
 		if r.DefinedTags != nil {
@@ -151,6 +150,10 @@ func (s *CoreSecurityListsDataSourceCrud) SetData() error {
 
 		if r.TimeCreated != nil {
 			securityList["time_created"] = r.TimeCreated.String()
+		}
+
+		if r.VcnId != nil {
+			securityList["vcn_id"] = *r.VcnId
 		}
 
 		resources = append(resources, securityList)