+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for DBaaS -  ADB - Serverless Extreme Availibility
 - Support for Switchover action in autonomous database added
+- Support for datasource of `core` with optional `vcn_id`
 
 ## 3.84.0 (July 08, 2020)